Split scan cli command into separate file
